requirements.txt
torch>=2.4.0
torchvision>=0.19.0
tqdm==4.66.4
pikepdf==9.0.0
pillow==10.4.0
numpy==1.26.4
tensorboard==2.17.0
pyyaml==6.0.2